backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
materialdash==0.0.24
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
